# recipes/service/spec/dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "recipes API",
    "version": "1.0.0",
    "description": "API specification for the recipes service"
  },
  "paths": {
...
    "/recipes/{id}": {
      "get": {
        "summary": "Get one recipe. Access from the recipe: public recipes allowed without auth; private recipes require auth and membership in the recipe's collection.",
        "operationId": "getRecipe",
        "tags": [
          "no-auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe with current version; optionally includes recent notes and file list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "recipe",
                    "currentVersion"
                  ],
                  "properties": {
                    "recipe": {
                      "$ref": "#/components/schemas/recipe"
                    },
                    "currentVersion": {
                      "$ref": "#/components/schemas/recipe-version"
                    },
                    "notes": {
                      "type": "array",
                      "description": "Optional list of recent recipe notes.",
                      "items": {
                        "type": "object"
                      }
                    },
                    "files": {
                      "type": "array",
                      "description": "Optional list of recipe file info.",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - recipe is private and caller is not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - not a member of the recipe's collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update recipe metadata only (title, descriptions, isPublic). Admin only.",
        "operationId": "updateRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Fields to update; only provided fields are updated.",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Recipe title"
                  },
                  "subtitle": {
                    "type": "string",
                    "description": "Subtitle shown on menus and list views"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true,
                    "description": "Optional longer description shown on recipe detail"
                  },
                  "isPublic": {
                    "type": "boolean",
                    "description": "Whether the recipe is visible to non-admin users"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated recipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/recipe"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete recipe (and handle versions/notes/files per policy). Admin only.",
        "operationId": "deleteRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          }
        ],
        "responses": {
          "204": {
            "description": "Recipe deleted."
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
...
    "/recipes/{id}/notes/{noteId}": {
      "put": {
        "summary": "Edit note; server sets everEdited. Admin only.",
        "operationId": "notesUpdateRecipes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Note id"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Note body and optional recipe version link.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "body"
                ],
                "properties": {
                  "body": {
                    "$ref": "#/components/schemas/body"
                  },
                  "recipeVersionId": {
                    "$ref": "#/components/schemas/recipeVersionId"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated RecipeNote (everEdited set by server when applicable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/recipe-note"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a note on a recipe. Admin only.",
        "operationId": "notesDeleteRecipes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Note id"
          }
        ],
        "responses": {
          "204": {
            "description": "Note deleted."
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
...
    "/recipes/{id}/notes/{noteId}/files": {
      "get": {
        "summary": "List all files for a recipe note.",
        "operationId": "notesFilesListRecipes",
        "tags": [
          "no-auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe note id"
          }
        ],
        "responses": {
          "200": {
            "description": "All files for the note.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/recipe-note-file-info"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload a new file for a recipe note (multiple files allowed). Admin only.",
        "operationId": "notesFilesUploadRecipes",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe id"
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Recipe note id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The file to upload (stored in Azure)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created recipe note file metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/recipe-note-file-info"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid auth headers, or not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user does not have required privileges (admin only).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "recipe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the recipe",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "title": {
            "type": "string",
            "description": "Recipe title",
            "example": "Classic Chocolate Chip Cookies"
          },
          "subtitle": {
            "type": "string",
            "description": "Subtitle shown on menus and list views",
            "example": "Crispy edges, chewy centers"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Optional longer description shown on recipe detail (not on menu)",
            "example": "A crowd-pleasing recipe that works every time. Best with room-temperature butter."
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether the recipe is visible to non-admin users",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User id of the creator",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the recipe was created",
            "example": "2023-01-15T14:30:00Z"
          },
          "updatedBy": {
            "type": "string",
            "description": "User id of the last updater",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the recipe was last updated",
            "example": "2023-02-01T09:00:00Z"
          },
          "currentVersionId": {
            "type": "string",
            "description": "Optional id of the current/latest version for display",
            "example": "b2c3d4e5-e89b-12d3-a456-426614174002"
          }
        },
        "required": [
          "id",
          "title",
          "subtitle",
          "isPublic",
          "createdBy",
          "createdAt",
          "updatedBy",
          "updatedAt"
        ]
      },
      "error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "A short, machine-readable error code, for when HTTP status codes are not sufficient."
          },
          "message": {
            "type": "string",
            "description": "A human-readable description of the error.",
            "example": "The requested resource could not be found."
          }
        }
      },
...
      "recipe-version": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for this recipe version",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "recipeId": {
            "type": "string",
            "description": "Id of the recipe this version belongs to",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "content": {
            "type": "object",
            "description": "Version body — structured ingredients and markdown instructions",
            "properties": {
              "ingredients": {
                "type": "array",
                "description": "List of ingredients; quantity and unit are strings (e.g. \"2+1/3\", \"cups\")",
                "items": {
                  "type": "object",
                  "required": [
                    "name",
                    "quantity",
                    "unit"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "Ingredient name",
                      "example": "All-purpose flour"
                    },
                    "quantity": {
                      "type": "string",
                      "description": "Quantity as string to support values like \"1/3\", \"2+1/3\"",
                      "example": "2+1/3"
                    },
                    "unit": {
                      "type": "string",
                      "description": "Unit of measure (unrestricted string)",
                      "example": "cups"
                    }
                  }
                },
                "example": [
                  {
                    "name": "All-purpose flour",
                    "quantity": "2",
                    "unit": "cups"
                  },
                  {
                    "name": "Butter",
                    "quantity": "1/2",
                    "unit": "cup"
                  }
                ]
              },
              "instructionsMarkdown": {
                "type": "string",
                "description": "Recipe instructions in markdown",
                "example": "1. Preheat oven to 350°F.\n2. Mix dry ingredients..."
              }
            },
            "required": [
              "ingredients",
              "instructionsMarkdown"
            ],
            "example": {
              "ingredients": [
                {
                  "name": "All-purpose flour",
                  "quantity": "2",
                  "unit": "cups"
                }
              ],
              "instructionsMarkdown": "1. Preheat oven to 350°F.\n2. Mix dry ingredients."
            }
          },
          "isLatest": {
            "type": "boolean",
            "description": "Whether this version is the current one for the recipe (only one per recipe is true)",
            "example": true
          },
          "createdBy": {
            "type": "string",
            "description": "User id of the creator of this version",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this version was created",
            "example": "2023-01-15T14:30:00Z"
          }
        },
        "required": [
          "id",
          "recipeId",
          "content",
          "isLatest",
          "createdBy",
          "createdAt"
        ]
      },
      "recipe-note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the note",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "recipeId": {
            "type": "string",
            "description": "Id of the recipe this note belongs to",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "recipeVersionId": {
            "type": "string",
            "nullable": true,
            "description": "Optional id of the recipe version this note is linked to (e.g. when written with a change)",
            "example": "b2c3d4e5-e89b-12d3-a456-426614174002"
          },
          "body": {
            "type": "string",
            "description": "Note content (how it went, what to try next)",
            "example": "Reduced sugar by 1/4 cup; next time try brown butter."
          },
          "everEdited": {
            "type": "boolean",
            "description": "Whether the note has ever been edited after creation (set to true on first update)",
            "example": false
          },
          "createdBy": {
            "type": "string",
            "description": "User id of the creator",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the note was created",
            "example": "2023-01-15T14:30:00Z"
          },
          "updatedBy": {
            "type": "string",
            "description": "User id of the last updater",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the note was last updated",
            "example": "2023-02-01T09:00:00Z"
          }
        },
        "required": [
          "id",
          "recipeId",
          "body",
          "everEdited",
          "createdBy",
          "createdAt",
          "updatedBy",
          "updatedAt"
        ]
      },
      "body": {
        "type": "string",
        "description": "Note content (how it went, what to try next)",
        "example": "Reduced sugar by 1/4 cup; next time try brown butter."
      },
      "recipeVersionId": {
        "type": "string",
        "nullable": true,
        "description": "Optional id of the recipe version this note is linked to (e.g. when written with a change)",
        "example": "b2c3d4e5-e89b-12d3-a456-426614174002"
      },
...
      "recipe-note-file-info": {
        "type": "object",
        "description": "Metadata for one file attached to a recipe note. Uses SAF file metadata (blob_name, file_original_name, mimetype, size, created_at, updated_at).",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the file record",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "recipeNoteId": {
            "type": "string",
            "description": "Id of the recipe note this file is attached to",
            "example": "b2c3d4e5-f90c-23e4-b567-537725285002"
          },
          "blobName": {
            "type": "string",
            "description": "Storage blob name (Azure) for the file",
            "example": "recipe-notes/b2c3d4e5-f90c-23e4-b567-537725285002/123e4567-e89b-12d3-a456-426614174000.pdf"
          },
          "fileOriginalName": {
            "type": "string",
            "description": "Original filename as provided on upload",
            "example": "attempt-photo.pdf"
          },
          "mimetype": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "application/pdf"
          },
          "size": {
            "type": "integer",
            "minimum": 0,
            "description": "File size in bytes",
            "example": 102400
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the file was uploaded",
            "example": "2023-01-15T14:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the file metadata was last updated",
            "example": "2023-01-15T14:30:00Z"
          },
          "uploadedBy": {
            "type": "string",
            "nullable": true,
            "description": "User id of the uploader (optional)",
            "example": "a1b2c3d4-e89b-12d3-a456-426614174001"
          },
          "downloadUrl": {
            "type": "string",
            "format": "uri",
            "description": "Full URL to download or view the file (API base + /recipes/:id/notes/:noteId/files/:fileId/blob)",
            "example": "https://api.recipes.example.com/recipes/a1b2c3d4-e89b-12d3-a456-426614174001/notes/b2c3d4e5-f90c-23e4-b567-537725285002/files/123e4567-e89b-12d3-a456-426614174000/blob"
          }
        },
        "required": [
          "id",
          "recipeNoteId",
          "blobName",
          "fileOriginalName",
          "mimetype",
          "size",
          "createdAt",
          "updatedAt",
          "downloadUrl"
        ]
      },
...
    }
  }
}